Revert "Moved components to a new admin folder"
