+++ app/admin/page.tsx
@@ -5,12 +5,12 @@ import { redirect } from "next/navigation";
 import Link from "next/link";
 import { Users, Edit, Sword, Car, X, BookOpen } from "lucide-react";
 import { useState, useEffect } from "react";
-import { AdminCreateFighterTypeModal } from "@/components/admin/admin-create-fighter-type";
-import { AdminEditFighterTypeModal } from "@/components/admin/admin-edit-fighter-type";
-import { AdminCreateEquipmentModal } from "@/components/admin/admin-create-equipment";
-import { AdminEditEquipmentModal } from "@/components/admin/admin-edit-equipment";
-import { AdminCreateSkillModal } from "@/components/admin/admin-create-skill";
-import { AdminEditSkillModal } from "@/components/admin/admin-edit-skill";
+import { AdminCreateFighterTypeModal } from "@/components/ui/admin-create-fighter-type";
+import { AdminEditFighterTypeModal } from "@/components/ui/admin-edit-fighter-type";
+import { AdminCreateEquipmentModal } from "@/components/ui/admin-create-equipment";
+import { AdminEditEquipmentModal } from "@/components/ui/admin-edit-equipment";
+import { AdminCreateSkillModal } from "@/components/ui/admin-create-skill";
+import { AdminEditSkillModal } from "@/components/ui/admin-edit-skill";
 import Modal from "@/components/modal";
 import { useToast } from "@/components/ui/use-toast";
 import { ToastProvider } from "@/components/ui/toast";
