+++ components/ui/admin-create-equipment.tsx
