+++ components/ui/admin-create-fighter-type.tsx
