+++ components/ui/admin-edit-equipment.tsx
