+++ components/ui/admin-edit-fighter-type.tsx
